+++ CHANGELOG.md
@@ -4,7 +4,7 @@ As this project is pre 1.0, breaking changes may happen for minor version bumps.
 
 ## Next version
 
-_Nothing, yet._
+- Added `base_offer_id` and `counter_offer_id` to the Trade object. These fields were introduced with Horizon v0.15.0 [#68](https://github.com/Synesso/scala-stellar-sdk/issues/68)
 
 
 ## 0.6.2